Refactor application types and detained license workflow
Introduce enums for application types and license classes, replacing string-based lookups with strongly-typed dictionaries. Refactor detained license release workflow: update UI to show detailed info, make release operation async, and improve fee calculation. Change FineFees to decimal for accuracy. Add repository method to fetch detained license by license ID. Update controller and repository logic for streamlined operations. Fix SQL table name and clarify UI labels. Improves type safety, maintainability, and user experience.

+++ DVLD/Pop-Ups/Detain_License.cs
@@ -100,7 +100,7 @@ namespace DVLD.Pop_Ups
                 {
                     LicenseID = _licenseId,
                     DetainDate = DateTime.Now,
-                    FineFees = float.Parse(txt_fine_fees_detain.Text.Trim()),
+                    FineFees = decimal.Parse(txt_fine_fees_detain.Text.Trim()),
                     CreatedByUserID = Session.CurrentUserProvider.CurrentUser.UserId,
                     IsReleased = false
                 };
